Move home viewModel to home package
