+++ app/src/main/java/com/demo/developer/deraesw/demomoviewes/ui/home/HomeFragment.kt
@@ -13,7 +13,6 @@ import com.demo.developer.deraesw.demomoviewes.R
 import com.demo.developer.deraesw.demomoviewes.adapter.HomePageAdapter
 import com.demo.developer.deraesw.demomoviewes.data.model.AccountData
 import com.demo.developer.deraesw.demomoviewes.databinding.FragmentHomeBinding
-import com.demo.developer.deraesw.demomoviewes.ui.HomeViewModel
 import com.demo.developer.deraesw.demomoviewes.utils.AppTools
 import com.google.android.material.tabs.TabLayoutMediator
 import dagger.hilt.android.AndroidEntryPoint
